// src/app/core/shared/models/cart-item-customer.model.ts
export interface cartItemCustomer{
    cartitem_id: number | any;
    product_id: number | any;
    cust_id: number | any;
    order_id: number | any;
    seller_id: number | any;
    product_name: string | any;
    price: number | any;
    stock: number | any;
    quantity: number | any;
    total: string | any;
    path: string | any;
}